requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
PyPDF2
openai
python-dotenv
requests
